+++ client/src/App.js
@@ -3,6 +3,7 @@ import { Route, Switch } from 'react-router-dom';
 import HomePage from './pages/HomePage';
 import SampleFormPage from './pages/SampleFormPage';
 import MapToolPage from './pages/MapToolPage';
+import ApiTestPage from './pages/ApiTestPage';
 
 import './App.css';
 
@@ -14,6 +15,7 @@ class App extends Component {
           <Route exact path="/" component={HomePage}/>
           <Route path="/sample_form" component={SampleFormPage}/>
           <Route path="/map_tool" component={MapToolPage}/>
+          <Route path="/api_test" component={ApiTestPage}/>
         </Switch>
       </div>
     );