+++ server/server.js
@@ -1,0 +1,19 @@
+const express = require('express');
+
+const app = express();
+const port = process.env.PORT || 3001;
+
+app.get('/api/hello', (req, res) => {
+  res.send({ express: 'Hello From Express' });
+});
+
+if (process.env.NODE_ENV === 'production') {
+  app.use(express.static(path.join(__dirname, 'client/build')));
+
+  // Client-side routing
+  app.get('*', function(req, res) {
+    res.sendFile(path.join(__dirname, 'client/build', 'index.html'));
+  });
+}
+
+app.listen(port, () => console.log(`Listening on port ${port}`));